requirements.txt
aenum==2.2.6
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
Django==4.2.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-configurations==2.5
django-cors-headers==4.3.0
django-filter==23.3
django-rest-swagger==2.2.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
fbchat==1.9.7
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
Naked==0.1.32
openapi-codec==1.3.2
packaging==23.2
paho-mqtt==1.6.1
prompt-toolkit==3.0.39
pycryptodome==3.19.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
shellescape==3.8.1
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
whitenoise==6.6.0
